Install jest, supertest and separate server files into 2: app.js and server.js
