+++ package.json
@@ -35,6 +35,8 @@
     "eslint-plugin-import": "^2.12.0",
     "eslint-plugin-jsx-a11y": "^6.0.3",
     "eslint-plugin-react": "^7.8.2",
+    "jest": "^23.1.0",
+    "supertest": "^3.1.0",
     "webpack": "^4.10.2",
     "webpack-cli": "^3.0.1"
   }
